requirements.txt
streamlit==1.55.0
pandas==2.3.3
numpy==2.4.3
plotly==6.6.0
scikit-learn==1.8.0
holidays==0.93
requests==2.32.5
opencv-python
ultralytics
python-dotenv